+++ src/sphere.rs
@@ -7,12 +7,12 @@ use crate::vec3::Vec3;
 #[derive(Clone)]
 pub struct Sphere {
     pub center: Vec3,
-    pub radius: f64,
+    pub radius: f32,
     pub material: Material,
 }
 
 impl Hitable for Sphere {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let oc = r.origin - self.center;
         let a = Vec3::dot(r.direction, r.direction);
         let b = Vec3::dot(oc, r.direction);
@@ -64,10 +64,10 @@ impl Hitable for Sphere {
     }
 }
 
-fn sphere_uv(p: &Vec3) -> (f64, f64) {
+fn sphere_uv(p: &Vec3) -> (f32, f32) {
     let phi = p.z().atan2(p.x());
     let theta = p.y().asin();
-    let u = 1.0 - (phi + std::f64::consts::PI) / (2.0 * std::f64::consts::PI);
-    let v = (theta + std::f64::consts::FRAC_PI_2) / std::f64::consts::PI;
+    let u = 1.0 - (phi + std::f32::consts::PI) / (2.0 * std::f32::consts::PI);
+    let v = (theta + std::f32::consts::FRAC_PI_2) / std::f32::consts::PI;
     (u, v)
 }
